+++ super_admin/app/Orchid/Screens/EditLimoGroupScreen.php
@@ -171,11 +171,11 @@ class EditLimoGroupScreen extends Screen
         //check if user is part of a limo group
         $user_limo_group = LimoGroupMember::where('invitee_user_id', $fields['creator_user_id'])->where('status', 1)->first();
 
-        if($owned_limo_group->id != $limoGroup->id){
+        if($owned_limo_group != null && $owned_limo_group->id != $limoGroup->id){
             //delete the old limo group
             $owned_limo_group->delete();
 
-        } elseif($user_limo_group->limo_group_id != $limoGroup->id){
+        } elseif($user_limo_group != null && $user_limo_group->limo_group_id != $limoGroup->id){
             //remove them as a limo group member from old limo group
             $user_limo_group->delete();
         }
